:hammer: Put `styles` in same folder.
